+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change default database to `sqlite`.
 - Change the `esplora-reqwest` feature to always use async mode
 - Change rpc `--skip-blocks` option to `--start-time` which specifies time initial sync will start scanning from.
+- Add new `bdk-cli node <command> [<args>]` to control the backend node deployed by `regtest-*` features.
+- Add an integration testing framework in `src/tests/integration.rs`. This framework uses the `regtest-*` feature to run automated testing with bdk-cli.
 
 ## [0.5.0]